Nullable reference type enabled.
This support forced us to upgrade to netstandard2.1 in order to
reference CSemVer that is now netstandard2.1 since it uses nullable
attributes like [AllowNull].
This is a Breaking change and the public API has changed, for instance
CIReleaseInfo.BaseTag is no a SVersion instead of a CSVersion.
The edge case when a CI build has no base tag now generates an
InformationalVersion with the commit sha and date.